refactor: separate WebSocket connection from game logic
